# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
paths:
...
  /api/teams:
    get:
      tags:
      - team-controller
      operationId: list
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      - name: includeDeleted
        in: query
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseTeamResponse"
    post:
      tags:
      - team-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamResponse"
  /api/teams/{teamId}/members:
    get:
      tags:
      - team-member-controller
      operationId: list_1
      parameters:
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMemberResponse"
    post:
      tags:
      - team-member-controller
      operationId: add
      parameters:
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamMemberResponse"
  /api/teams/batch:
    post:
      tags:
      - batch-import-controller
      operationId: importTeams
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamBatchImportRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamBatchImportResponse"
...
  /api/contests/{contestId}/teams:
    get:
      tags:
      - contest-team-controller
      operationId: list_5
      parameters:
      - name: contestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseContestTeamResponse"
    post:
      tags:
      - contest-team-controller
      operationId: assign
      parameters:
      - name: contestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContestTeamAssignmentRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContestTeamResponse"
...
  /api/teams/{teamId}/members/{memberId}:
    delete:
      tags:
      - team-member-controller
      operationId: remove
      parameters:
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: memberId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
    patch:
      tags:
      - team-member-controller
      operationId: update_1
      parameters:
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: memberId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberPatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamMemberResponse"
  /api/teams/{id}:
    get:
      tags:
      - team-controller
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamResponse"
    delete:
      tags:
      - team-controller
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
    patch:
      tags:
      - team-controller
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamPatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamResponse"
...
  /api/contests/{contestId}/teams/{teamId}:
    delete:
      tags:
      - contest-team-controller
      operationId: unassign
      parameters:
      - name: contestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
components:
  schemas:
...
    TeamRequest:
      required:
      - name
      type: object
      properties:
        name:
          type: string
        school:
          type: string
        seatNo:
          type: string
        groupName:
          type: string
        star:
          type: boolean
        username:
          type: string
        initialPassword:
          type: string
    TeamResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        school:
          type: string
        seatNo:
          type: string
        groupName:
          type: string
        star:
          type: boolean
        deletedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TeamMemberRequest:
      required:
      - name
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        roleName:
          type: string
    TeamMemberResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        teamId:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        roleName:
          type: string
        createdAt:
          type: string
          format: date-time
    TeamBatchImportRequest:
      type: object
      properties:
        teams:
          maxItems: 1000
          minItems: 1
          type: array
          items:
            $ref: "#/components/schemas/TeamRequest"
        contestId:
          type: integer
          format: int64
        participationType:
          type: string
          enum:
          - OFFICIAL
          - STAR
          - PRACTICE
        idempotencyKey:
          type: string
    RowResult:
      type: object
      properties:
        index:
          type: integer
          format: int32
        status:
          type: string
        teamId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        username:
          type: string
        initialPassword:
          type: string
        error:
          type: string
    TeamBatchImportResponse:
      type: object
      properties:
        batchId:
          type: string
        totalRequested:
          type: integer
          format: int32
        succeeded:
          type: integer
          format: int32
        failed:
          type: integer
          format: int32
        results:
          type: array
          items:
            $ref: "#/components/schemas/RowResult"
...
    ContestTeamAssignmentRequest:
      type: object
      properties:
        teamId:
          type: integer
          format: int64
        participationType:
          type: string
          enum:
          - OFFICIAL
          - STAR
          - PRACTICE
        groupName:
          type: string
    ContestTeamResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        contestId:
          type: integer
          format: int64
        teamId:
          type: integer
          format: int64
        teamName:
          type: string
        participationType:
          type: string
          enum:
          - OFFICIAL
          - STAR
          - PRACTICE
        groupName:
          type: string
        createdAt:
          type: string
          format: date-time
...
    TeamMemberPatchRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        roleName:
          type: string
    TeamPatchRequest:
      type: object
      properties:
        name:
          type: string
        school:
          type: string
        seatNo:
          type: string
        groupName:
          type: string
        star:
          type: boolean
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageResponseTeamResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/TeamResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    PageResponseContestTeamResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ContestTeamResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...